Refactor components and file structure
